We added update professor action

+++ newproject/Software2 project/Models/ProfessorModel.cs
@@ -36,7 +36,7 @@ namespace Software2_project.Models
 
         [Required]
         [Display(Name = "Salary")]
-        public float salary { get; set; }
+        public double salary { get; set; }
 
         [Required]
         [EmailAddress]
